--- operator/pkg/helm/testdata/addons/invalid/b/Chart.yaml
@@ -1,8 +1,0 @@
-apiVersion: v1
-description: A Helm chart for Kubernetes
-name: addon
-version: 1.1.0
-appVersion: 1.1.0
-tillerVersion: ">=2.7.2"
-keywords:
-  - istio-addon
