--- operator/pkg/helm/vfs_renderer_test.go
@@ -1,37 +1,0 @@
-// Copyright Istio Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package helm
-
-import "testing"
-
-func TestStripPrefix(t *testing.T) {
-	cases := []struct {
-		in     string
-		prefix string
-		out    string
-	}{
-		{"foo/bar", "foo", "bar"},
-		{"a/b/c/d", "a/b/c", "d"},
-		{"a/b/c/d", "a/b", "c/d"},
-	}
-	for _, tt := range cases {
-		t.Run(tt.in+"~"+tt.prefix, func(t *testing.T) {
-			got := stripPrefix(tt.in, tt.prefix)
-			if got != tt.out {
-				t.Fatalf("Wanted %v got %v", tt.out, got)
-			}
-		})
-	}
-}
